+++ backend/src/main/java/com/minibeit/post/domain/repository/PostRepositoryImpl.java
@@ -1,11 +1,11 @@
 package com.minibeit.post.domain.repository;
 
-import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.post.domain.Payment;
 import com.minibeit.post.domain.Post;
 import com.minibeit.post.domain.PostStatus;
-import com.minibeit.post.dto.PostResponse;
-import com.minibeit.post.dto.QPostResponse_GetMyApplyList;
+import com.minibeit.post.service.dto.PostResponse;
+import com.minibeit.post.service.dto.QPostResponse_GetMyApplyList;
+import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.user.domain.User;
 import com.querydsl.core.QueryResults;
 import com.querydsl.core.types.dsl.BooleanExpression;
