+++ backend/src/main/java/com/minibeit/post/service/PostByBusinessService.java
@@ -15,8 +15,8 @@ import com.minibeit.post.domain.RejectPost;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
 import com.minibeit.post.domain.repository.RejectPostRepository;
-import com.minibeit.post.dto.PostRequest;
-import com.minibeit.post.dto.PostResponse;
+import com.minibeit.post.service.dto.PostRequest;
+import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.exception.PostNotFoundException;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.postapplicant.domain.PostApplicant;
