+++ backend/src/main/java/com/minibeit/post/service/PostService.java
@@ -5,7 +5,7 @@ import com.minibeit.post.domain.*;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostLikeRepository;
 import com.minibeit.post.domain.repository.PostRepository;
-import com.minibeit.post.dto.PostResponse;
+import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.exception.PostNotFoundException;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.security.userdetails.CustomUserDetails;
