+++ backend/src/main/java/com/minibeit/postapplicant/service/PostApplicantByBusinessService.java
@@ -6,10 +6,10 @@ import com.minibeit.post.domain.Post;
 import com.minibeit.post.domain.PostDoDate;
 import com.minibeit.post.domain.RejectPost;
 import com.minibeit.post.domain.repository.RejectPostRepository;
-import com.minibeit.post.dto.PostResponse;
+import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.PostPermissionCheck;
-import com.minibeit.post.service.exception.ExistedApplySameTimeException;
-import com.minibeit.post.service.exception.PostApplicantNotFoundException;
+import com.minibeit.postapplicant.service.exception.ExistedApplySameTimeException;
+import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
 import com.minibeit.post.service.exception.PostDoDateIsFullException;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.postapplicant.domain.PostApplicant;
