+++ backend/src/main/java/com/minibeit/postapplicant/service/PostApplicantService.java
@@ -11,8 +11,8 @@ import com.minibeit.postapplicant.domain.repository.PostApplicantRepository;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostLikeRepository;
 import com.minibeit.postapplicant.service.dto.PostApplicantResponse;
-import com.minibeit.post.service.exception.DuplicateApplyException;
-import com.minibeit.post.service.exception.PostApplicantNotFoundException;
+import com.minibeit.postapplicant.service.exception.DuplicateApplyException;
+import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
 import com.minibeit.post.service.exception.PostDoDateIsFullException;
 import com.minibeit.post.service.exception.PostDoDateNotFoundException;
 import com.minibeit.user.domain.User;
