+++ backend/src/test/java/com/minibeit/post/service/PostByBusinessServiceTest.java
@@ -17,11 +17,11 @@ import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostLikeRepository;
 import com.minibeit.post.domain.repository.PostRepository;
 import com.minibeit.post.domain.repository.RejectPostRepository;
-import com.minibeit.post.dto.PostDto;
-import com.minibeit.post.dto.PostRequest;
-import com.minibeit.post.dto.PostResponse;
-import com.minibeit.post.service.exception.ExistApprovedApplicantException;
-import com.minibeit.post.service.exception.PostApplicantNotFoundException;
+import com.minibeit.post.service.dto.PostDto;
+import com.minibeit.post.service.dto.PostRequest;
+import com.minibeit.post.service.dto.PostResponse;
+import com.minibeit.postapplicant.service.exception.ExistApprovedApplicantException;
+import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
 import com.minibeit.post.service.exception.PostNotFoundException;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.postapplicant.domain.PostApplicant;
