+++ backend/src/test/java/com/minibeit/post/service/PostFindServiceTest.java
@@ -18,9 +18,9 @@ import com.minibeit.post.domain.PostLike;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostLikeRepository;
 import com.minibeit.post.domain.repository.PostRepository;
-import com.minibeit.post.dto.PostDto;
-import com.minibeit.post.dto.PostRequest;
-import com.minibeit.post.dto.PostResponse;
+import com.minibeit.post.service.dto.PostDto;
+import com.minibeit.post.service.dto.PostRequest;
+import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.exception.PostNotFoundException;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.postapplicant.domain.PostApplicant;
