+++ backend/src/test/java/com/minibeit/post/service/RejectPostServiceTest.java
@@ -4,7 +4,7 @@ import com.minibeit.ServiceIntegrationTest;
 import com.minibeit.common.dto.PageDto;
 import com.minibeit.post.domain.RejectPost;
 import com.minibeit.post.domain.repository.RejectPostRepository;
-import com.minibeit.post.dto.RejectPostResponse;
+import com.minibeit.post.service.dto.RejectPostResponse;
 import com.minibeit.user.domain.Role;
 import com.minibeit.user.domain.SignupProvider;
 import com.minibeit.user.domain.User;
