+++ backend/src/test/java/com/minibeit/post/web/PostByBusinessControllerTest.java
@@ -7,9 +7,9 @@ import com.minibeit.post.domain.Payment;
 import com.minibeit.post.domain.Post;
 import com.minibeit.post.domain.PostDoDate;
 import com.minibeit.file.domain.PostFile;
-import com.minibeit.post.dto.PostDto;
-import com.minibeit.post.dto.PostRequest;
-import com.minibeit.post.dto.PostResponse;
+import com.minibeit.post.service.dto.PostDto;
+import com.minibeit.post.service.dto.PostRequest;
+import com.minibeit.post.service.dto.PostResponse;
 import com.minibeit.post.service.PostByBusinessService;
 import com.minibeit.school.domain.School;
 import com.minibeit.user.domain.User;
