+++ backend/src/test/java/com/minibeit/postapplicant/service/PostApplicantByBusinessFindServiceTest.java
@@ -11,11 +11,10 @@ import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
 import com.minibeit.postapplicant.service.dto.PostApplicantDto;
 import com.minibeit.postapplicant.service.dto.PostApplicantResponse;
-import com.minibeit.post.dto.PostDto;
-import com.minibeit.post.dto.PostRequest;
+import com.minibeit.post.service.dto.PostDto;
+import com.minibeit.post.service.dto.PostRequest;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.postapplicant.domain.PostApplicant;
-import com.minibeit.postapplicant.service.PostApplicantByBusinessService;
 import com.minibeit.school.domain.School;
 import com.minibeit.school.domain.SchoolRepository;
 import com.minibeit.user.domain.Role;
@@ -32,7 +31,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
