+++ backend/src/test/java/com/minibeit/postapplicant/service/PostApplicantByBusinessServiceTest.java
@@ -13,10 +13,10 @@ import com.minibeit.post.domain.RejectPost;
 import com.minibeit.post.domain.repository.PostDoDateRepository;
 import com.minibeit.post.domain.repository.PostRepository;
 import com.minibeit.post.domain.repository.RejectPostRepository;
-import com.minibeit.post.dto.PostDto;
-import com.minibeit.post.dto.PostRequest;
-import com.minibeit.post.service.exception.ExistedApplySameTimeException;
-import com.minibeit.post.service.exception.PostApplicantNotFoundException;
+import com.minibeit.post.service.dto.PostDto;
+import com.minibeit.post.service.dto.PostRequest;
+import com.minibeit.postapplicant.service.exception.ExistedApplySameTimeException;
+import com.minibeit.postapplicant.service.exception.PostApplicantNotFoundException;
 import com.minibeit.post.service.exception.PostDoDateIsFullException;
 import com.minibeit.postapplicant.domain.ApplyStatus;
 import com.minibeit.postapplicant.domain.PostApplicant;
